Lots of changes - everything necessary to create a crude bot that just masses marines and attacks with them.
Still a lot of flaws in its behavior (the biggest being it builds too many Barracks and not enough Supply Depots).

Also filled out more common logic and fixed some issues with the map data (was not putting the bytes in an array that worked the way I thought it did).